+++ src/scattermind/system/executor/executor.py
@@ -1,18 +1,16 @@
-# Scattermind distributes computation of machine learning models.
 # Copyright (C) 2024 Josua Krause
 #
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 """Executor that computes the execution graph."""
 import time
 import traceback
@@ -142,7 +140,7 @@ class ExecutorManager(Module):
                     {
                         "name": "node",
                         "action": "unload",
-                        "target": node.get_name(),
+                        "target": node.get_qualified_name(queue_pool),
                     })
                 queue_pool.remove_node_listener(node, own_id)
                 node.unload(own_id)
@@ -151,7 +149,7 @@ class ExecutorManager(Module):
                 {
                     "name": "node",
                     "action": "load",
-                    "target": new_node.get_name(),
+                    "target": new_node.get_qualified_name(queue_pool),
                 })
             queue_pool.add_node_listener(new_node, own_id)
             new_node.load(own_id, roa)
@@ -161,7 +159,7 @@ class ExecutorManager(Module):
                 {
                     "name": "node",
                     "action": "load_done",
-                    "target": new_node.get_name(),
+                    "target": new_node.get_qualified_name(queue_pool),
                 })
         assert self._node is not None
         return self._node