+++ src/scattermind/system/readonly/loader.py
@@ -1,18 +1,16 @@
-# Scattermind distributes computation of machine learning models.
 # Copyright (C) 2024 Josua Krause
 #
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 """Create a readonly access."""
 from typing import Literal, TypedDict
 
@@ -22,6 +20,7 @@ from scattermind.system.readonly.access import ReadonlyAccess
 
 LocalReadonlyAccessModule = TypedDict('LocalReadonlyAccessModule', {
     "name": Literal["ram"],
+    "scratch": str,
 })
 """A RAM based readonly access. This can only be used locally and is used
 mostly for tests."""
@@ -53,5 +52,5 @@ def load_readonly_access(module: ReadonlyAccessModule) -> ReadonlyAccess:
         return plugin(**kwargs)
     if module["name"] == "ram":
         from scattermind.system.readonly.ram import RAMAccess
-        return RAMAccess()
+        return RAMAccess(module["scratch"])
     raise ValueError(f"unknown readonly access: {module['name']}")