+++ src/scattermind/system/readonly/ram.py
@@ -1,24 +1,23 @@
-# Scattermind distributes computation of machine learning models.
 # Copyright (C) 2024 Josua Krause
 #
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 """A RAM based readonly access. It has a writer interface initially fill the
 data. It is most commonly used for test cases."""
+import os
 import threading
 from io import BytesIO, SEEK_END, SEEK_SET
 
-from scattermind.system.base import L_EITHER, Locality
+from scattermind.system.base import L_EITHER, Locality, NodeId
 from scattermind.system.readonly.access import ReadonlyAccess
 from scattermind.system.readonly.writer import RoAWriter
 
@@ -26,9 +25,10 @@ from scattermind.system.readonly.writer import RoAWriter
 class RAMAccess(ReadonlyAccess[str], RoAWriter[str]):
     """A RAM based readonly access. It has a writer interface initially fill
     the data. It is most commonly used for test cases."""
-    def __init__(self) -> None:
+    def __init__(self, scratch: str) -> None:
         super().__init__()
         self._objs: dict[str, BytesIO] = {}
+        self._scratch_folder = scratch
         self._lock = threading.RLock()
 
     @staticmethod
@@ -60,3 +60,11 @@ class RAMAccess(ReadonlyAccess[str], RoAWriter[str]):
 
     def close(self, hnd: str) -> None:
         pass
+
+    def do_get_scratchspace(self, node_id: NodeId) -> str:
+        name = node_id.to_parseable()
+        prefix = name[:3]
+        postfix = name[3:]
+        path = os.path.join(self._scratch_folder, prefix, postfix)
+        os.makedirs(path, exist_ok=True)
+        return path