+++ test/test_util.py
@@ -1,19 +1,19 @@
-# Scattermind distributes computation of machine learning models.
 # Copyright (C) 2024 Josua Krause
 #
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 """Tests for the utility modules."""
+import os
+import tomllib
 from typing import Any
 
 import numpy as np
@@ -292,3 +292,16 @@ def test_invalid() -> None:
         mask_from_shape([1, 2, 3], [2, 1, 3])
     with pytest.raises(ValueError, match=r"invalid str from tensor"):
         tensor_to_str(create_tensor(np.array([240, 159, 152, 0]), dtype="int"))
+
+
+def test_version() -> None:
+    """Test the version field."""
+    import scattermind  # pylint: disable=import-outside-toplevel
+
+    fname = os.path.join(os.path.dirname(__file__), "../pyproject.toml")
+    with open(fname, "rb") as fin:
+        pyproject = tomllib.load(fin)
+    version = pyproject["project"]["version"]
+    versions = (version, f"{version}*")
+    assert scattermind.__version__ in versions
+    assert scattermind.version in versions